Driver Models

Driver Models

Behavior models for cars and pedestrians.

AutomotivePOMDPs.ConstantSpeedDawdling — Type.

Action type for the pedestrian model

source

AutomotivePOMDPs.set_direction! — Method.

set the lane exit to take at the next junction to reach the goal

source

AutomotivePOMDPs.update_priority! — Method.

Check if all the pedestrian have crossed

source

AutomotiveDrivingModels.track_longitudinal! — Method.

Compute acceleration according to IDM

source

AutomotivePOMDPs.get_dist_to_end — Method.

Return the distance to the end of the lane

source

AutomotivePOMDPs.stop_at_dist — Method.

Return the longitudinal acceleration to come to a stop at the distance headway following idm

source

AutomotivePOMDPs.stop_at_end — Method.

Return the longitudinal acceleration to come to a stop at the end of the current lane follows idm style braking

source

AutomotivePOMDPs.update_stop! — Method.

set the state stop! to true if veh is stopped at the intersection

source

AutomotivePOMDPs.grow_wait_list! — Method.

If other vehicles reaches the intersection before the ego vehicle they are added to the waitlist

source

AutomotivePOMDPs.is_at_intersection — Method.

return true if veh is before the intersection

source

AutomotivePOMDPs.tts — Method.

Robust function to compute the time to drive dist at speed v

source

AutomotivePOMDPs.ungrow_wait_list! — Method.

Remove from the wait list vehicles that are now exiting the intersection

source

AutomotivePOMDPs.update_priority! — Method.

Check if the ego car has priority or not

source

AutomotivePOMDPs.ttc_check — Method.

Check if the time to collision is above some threshold

source